+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/domain/collections/MapEntry.kt
@@ -1,4 +1,4 @@
-package com.windea.breezeframework.core.domain.collections.deprecated
+package com.windea.breezeframework.core.domain.collections
 
 data class MapEntry<K, V>(
 	override val key: K,