+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/types/CollectionTypes.kt
@@ -1,18 +1,5 @@
 package com.windea.breezeframework.core.types
 
-typealias MultiValueSet<T> = Set<List<T>>
-
-typealias MutableMultiValueSet<T> = MutableSet<MutableList<T>>
-
 typealias MultiValueMap<K, V> = Map<K, List<V>>
 
 typealias MutableMultiValueMap<K, V> = MutableMap<K, MutableList<V>>
-
-
-typealias Table<T> = Map<T, List<T>>
-
-typealias MutableTable<T> = MutableMap<T, MutableList<T>>
-
-
-
-